Add two-column mobile-friendly Hugo site.
This is a simple place to host WebGL demos and tutorials, we can prettify later. It will show up at:

    https://google.github.io/filament/

The hugo config specifies the output holder to be ../docs which is where our GitHub Pages site is located.

Note that this commit adds the *source* to the site, it does not publish the actual site. When we're ready to publish the actual site, we'll do:

```terminal
cd site ; hugo ; cd ..
git add docs ; git commit
```