Work around WebGL depth artifacts seen on some GPUs.
As a temporary measure, this disables the depth prepass for the web
platform. Something in our shader pipeline (e.g. glslang or spirv-opt)
is not honoring the "invariant" keyword.

This fixes the black flakes seen with the Intel HD Graphics 615 in
Pixelbook laptops.

As a next step, I will dig deeper to find the offending component.